+++ spring-data-eclipse-store/src/main/java/software/xdev/spring/data/eclipse/store/repository/config/EclipseStoreClientConfiguration.java
@@ -20,9 +20,12 @@ import org.eclipse.store.integrations.spring.boot.types.factories.EmbeddedStorag
 import org.eclipse.store.storage.embedded.types.EmbeddedStorageFoundation;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.transaction.TransactionManagerCustomizers;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.event.ContextClosedEvent;
+import org.springframework.context.event.EventListener;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.TransactionManager;
 
@@ -51,11 +54,11 @@ import software.xdev.spring.data.eclipse.store.transactions.EclipseStoreTransact
 })
 public abstract class EclipseStoreClientConfiguration implements EclipseStoreStorageFoundationProvider
 {
-	private final EclipseStoreProperties defaultEclipseStoreProperties;
-	private final EmbeddedStorageFoundationFactory defaultEclipseStoreProvider;
+	protected final EclipseStoreProperties defaultEclipseStoreProperties;
+	protected final EmbeddedStorageFoundationFactory defaultEclipseStoreProvider;
 	
-	private EclipseStoreStorage storageInstance;
-	private EclipseStoreTransactionManager transactionManager;
+	protected EclipseStoreStorage storageInstance;
+	protected EclipseStoreTransactionManager transactionManager;
 	
 	@Autowired
 	protected EclipseStoreClientConfiguration(
@@ -99,10 +102,9 @@ public abstract class EclipseStoreClientConfiguration implements EclipseStoreSto
 	public PlatformTransactionManager transactionManager(
 		final ObjectProvider<TransactionManagerCustomizers> transactionManagerCustomizers)
 	{
-		final EclipseStoreTransactionManager transactionManager = this.getTransactionManagerInstance();
-		transactionManagerCustomizers.ifAvailable((customizers) ->
-			customizers.customize((TransactionManager)transactionManager));
-		return transactionManager;
+		final EclipseStoreTransactionManager tm = this.getTransactionManagerInstance();
+		transactionManagerCustomizers.ifAvailable(customizers -> customizers.customize((TransactionManager)tm));
+		return tm;
 	}
 	
 	public EclipseStoreTransactionManager getTransactionManagerInstance()
@@ -113,4 +115,71 @@ public abstract class EclipseStoreClientConfiguration implements EclipseStoreSto
 		}
 		return this.transactionManager;
 	}
+	
+	// region On context closed shutdown storage
+	
+	@Value("${spring-data-eclipse-store.context-close-shutdown-storage.enabled:true}")
+	protected boolean contextCloseShutdownStorageEnabled;
+	
+	@Value("${spring-data-eclipse-store.context-close-shutdown-storage.only-when-dev-tools:true}")
+	protected boolean contextCloseShutdownStorageOnlyWhenDevTools;
+	
+	/**
+	 * Upstream value from Spring Boot DevTools.
+	 *
+	 * @see org.springframework.boot.devtools.autoconfigure.DevToolsProperties.Restart
+	 */
+	@Value("${spring.devtools.restart.enabled:true}")
+	protected boolean springDevtoolsRestartEnabled;
+	
+	protected boolean shouldShutdownStorageOnContextClosed()
+	{
+		// Did the user disable support for this?
+		if(!this.contextCloseShutdownStorageEnabled)
+		{
+			return false;
+		}
+		
+		// Always or only for DevTools?
+		if(!this.contextCloseShutdownStorageOnlyWhenDevTools)
+		{
+			return true;
+		}
+		
+		// Spring DevTools loaded?
+		try
+		{
+			Class.forName("org.springframework.boot.devtools.autoconfigure.DevToolsProperties");
+		}
+		catch(final ClassNotFoundException e)
+		{
+			return false;
+		}
+		
+		// Spring Boot DevTools Restart enabled?
+		return this.springDevtoolsRestartEnabled;
+	}
+	
+	/**
+	 * Invoked when the application is "shut down" - or parts of it during a DevTools restart.
+	 * <p>
+	 * Shuts down the storage when it's present and {@link #shouldShutdownStorageOnContextClosed()} is
+	 * <code>true</code>
+	 * </p>
+	 *
+	 * <p>
+	 * This is required for the DevTools restart as it otherwise crashes with <code>StorageExceptionInitialization:
+	 * Active storage for ... already exists</code>
+	 * </p>
+	 */
+	@EventListener
+	public void shutdownStorageOnContextClosed(final ContextClosedEvent event)
+	{
+		if(this.storageInstance != null && this.shouldShutdownStorageOnContextClosed())
+		{
+			this.storageInstance.stop();
+		}
+	}
+	
+	// endregion
 }